pyproject.toml
[tool.poetry]
name = "lucy"
version = "0.1.0"
description = ""
authors = ["Vitaliy Parytskyi <v.parytskyi@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10.6, <3.12"
transformers = {git = "https://github.com/huggingface/transformers.git", rev = "main"}
starlette = "^0.21.0"
uvicorn = "^0.19.0"
sentence-transformers = "^2.2.2"
matplotlib = "^3.6.1"
pandas = "^1.5.1"
ctranslate2 = "^2.24.0"
fairseq = "^0.12.2"
mapply = "^0.1.16"
datasets = "^2.6.1"
googletrans = "^3.0.0"
gunicorn = "^20.1.0"
tensorboardX = "^2.5.1"
SudachiPy = "^0.6.6"
SudachiDict-core = "^20221021"
evaluate = "^0.3.0"
openpyxl = "^3.0.10"
ipython = "^8.6.0"
python-slugify = "^6.1.2"
polib = "^1.1.1"
pydantic = "^1.10.2"
beautifulsoup4 = "^4.11.1"
prettytable = "^3.5.0"
lupa = "^1.13"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
